feat: WIP refactor typeorm config to enable migrations
